+++ acceptance/support.tf
@@ -7,12 +7,12 @@ locals {
       {
         name = "var_one"
         ip   = "127.0.0.1"
-        vars = { "ansible_connection" = "local" }
+        vars = {}
       },
       {
         name = "var_other_one"
         ip   = "127.0.0.1"
-        vars = { "ansible_connection" = "local" }
+        vars = {}
       }
     ],
     "group_two" = [
@@ -28,12 +28,12 @@ locals {
   instances_aws = {
     "aws_one" = {
       id         = "i-1234567890"
-      tags       = { "Name" = "aws_one", "foo" = "bar", "ansible_connection" = "local" }
+      tags       = { "Name" = "aws_one", "foo" = "bar", "baz" = "bat" }
       private_ip = "127.0.0.1"
     },
     "aws_two" = {
       id         = "i-0987654321"
-      tags       = { "ansible_connection" = "local" }
+      tags       = {}
       private_ip = "127.0.0.1"
     }
   }
@@ -42,12 +42,12 @@ locals {
   instances_gcp = {
     "gcp_one" = {
       name              = "gcp_one"
-      tags              = ["Name=gcp_one", "foo=bar", "ansible_connection=local", "othertag", "not-host-var"]
+      tags              = ["Name=gcp_one", "foo=bar", "othertag", "not-host-var"]
       network_interface = [{ network_ip = "127.0.0.1" }]
     },
     "gcp_two" = {
       name              = "gcp_two"
-      tags              = ["ansible_connection=local"]
+      tags              = []
       network_interface = [{ network_ip = "127.0.0.1" }]
     }
   }
@@ -57,7 +57,7 @@ locals {
     "azr_one" = {
       id                 = "1234567890abcdefg"
       name               = "azr_one"
-      tags               = { "foo" = "bar", "ansible_connection" = "local" }
+      tags               = { "foo" = "bar", "baz" = "bat" }
       private_ip_address = "127.0.0.1"
       admin_ssh_key      = [{ "username" = "not_admin" }]
       admin_username     = "administrator"
@@ -65,7 +65,7 @@ locals {
     "azr_two" = {
       id                 = "gfedcba0987654321"
       name               = "azr_two"
-      tags               = { "ansible_connection" = "local" }
+      tags               = {}
       private_ip_address = "127.0.0.1"
       admin_username     = "administrator"
     }
@@ -76,14 +76,23 @@ locals {
     "vsp_one" = {
       name               = "vsp_one"
       default_ip_address = "127.0.0.1"
-      vapp               = [{ "properties" = { "foo" = "bar", "ansible_connection" = "local" } }]
+      vapp               = [{ "properties" = { "foo" = "bar", "baz" = "bat" } }]
     },
     "vsp_two" = {
       name               = "vsp_two"
       default_ip_address = "127.0.0.1"
-      vapp               = [{ "properties" = { "ansible_connection" = "local" } }]
+      vapp               = [{ "properties" = {} }]
       clone              = [{ "customize" = [{ "windows_options" = [{ "full_name" = "not_administrator" }] }] }]
     }
+  }
+
+  # mock group vars
+  group_vars = {
+    "group_one" = { "ansible_connection" = "local" },
+    "aws"       = { "ansible_connection" = "local" },
+    "gcp"       = { "ansible_connection" = "local" },
+    "azr"       = { "ansible_connection" = "local" },
+    "vsp"       = { "ansible_connection" = "local" }
   }
 }