+++ variables.tf
@@ -1,3 +1,4 @@
+# inventory formatting vars
 variable "formats" {
   type        = set(string)
   default     = []
@@ -9,6 +10,13 @@ variable "formats" {
   }
 }
 
+variable "prefix" {
+  type        = string
+  default     = ""
+  description = "A prefix to prepend to the name of the output inventory files. For example: the INI inventory will be named '<prefix>inventory.ini'."
+}
+
+# instance vars
 variable "instances" {
   type = map(
     set(
@@ -47,8 +55,9 @@ variable "instances_vsp" {
   description = "The 'vsphere_virtual_machine.this' map of objects comprising multiple instances to populate the Ansible inventory file."
 }
 
-variable "prefix" {
-  type        = string
-  default     = ""
-  description = "A prefix to prepend to the name of the output inventory files. For example: the INI inventory will be named '<prefix>inventory.ini'."
+# group vars
+variable "group_vars" {
+  type        = map(any)
+  default     = {}
+  description = "The map of Ansible group variables. Each key in the map is the name of a group, and each value is the object representing the pairs of group variable names and values."
 }